requirements.txt
mkdocs
PyGithub
requests
tqdm
pyyaml
funcy
